+++ backend/src/main/java/com/ankursinha/backend/repo/StudentRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepo extends JpaRepository<Student, Long> {
+public interface StudentRepo extends JpaRepository<Student, Integer> {
     Student findByEmail(String email);
     boolean existsByRollNum(String rollNum);
     boolean existsByEmail(String email);
